requirements.txt

livekit-agents[openai]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
livekit-agents[groq]~=1.0
livekit-plugins-cartesia
livekit-plugins-deepgram
livekit-plugins-silero #voice activity detection
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-turn-detector
livekit-agents
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib